+++ src/TrueLayer/Payments/Model/CreatePaymentRequest.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Text.Json.Serialization;
 using OneOf;
 using TrueLayer.Payments.Model.AuthorizationFlow;
 using static TrueLayer.Payments.Model.PaymentMethod;
@@ -24,6 +25,7 @@ namespace TrueLayer.Payments.Model
         /// If provided, the start authorization flow endpoint does not need to be called</param>
         /// <param name="metadata">Add to the payment a list of custom key-value pairs as metadata</param>
         /// <param name="riskAssessment">The risk assessment and the payment_creditable webhook configuration.</param>
+        /// <param name="subMerchants">The details related to any applicable sub-merchants</param>
         public CreatePaymentRequest(
             long amountInMinor,
             string currency,
@@ -32,7 +34,8 @@ namespace TrueLayer.Payments.Model
             RelatedProducts? relatedProducts = null,
             StartAuthorizationFlowRequest? authorizationFlow = null,
             Dictionary<string, string>? metadata = null,
-            RiskAssessment? riskAssessment = null)
+            RiskAssessment? riskAssessment = null,
+            PaymentSubMerchants? subMerchants = null)
         {
             AmountInMinor = amountInMinor.GreaterThan(0, nameof(amountInMinor));
             Currency = currency.NotNullOrWhiteSpace(nameof(currency));
@@ -42,6 +45,7 @@ namespace TrueLayer.Payments.Model
             AuthorizationFlow = authorizationFlow;
             Metadata = metadata;
             RiskAssessment = riskAssessment;
+            SubMerchants = subMerchants;
         }
 
         /// <summary>
@@ -84,5 +88,11 @@ namespace TrueLayer.Payments.Model
         /// Gets the risk assessment configuration
         /// </summary>
         public RiskAssessment? RiskAssessment { get; }
+
+        /// <summary>
+        /// Gets the sub-merchants details
+        /// </summary>
+        [JsonPropertyName("sub_merchants")]
+        public PaymentSubMerchants? SubMerchants { get; }
     }
 }