+++ src/TrueLayer/Payments/Model/ListPaymentRefundsResponse.cs
@@ -3,6 +3,6 @@ using OneOf;
 
 namespace TrueLayer.Payments.Model;
 
-using RefundUnion = OneOf<RefundPending, RefundAuthorized>;
+using RefundUnion = OneOf<RefundPending, RefundAuthorized, RefundExecuted, RefundFailed>;
 
 public record ListPaymentRefundsResponse(List<RefundUnion> Items);